change int ot long

+++ src/main/java/utils/FactorialUtils.java
@@ -1,6 +1,7 @@
 package utils;
 
 import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 /**
  * Класс для работы с факториалами
@@ -13,7 +14,7 @@ public class FactorialUtils {
      * @return возвращает факториал значения
      * @throws Exception
      */
-    public static int calcFactorial(int num) throws Exception {
+    public static long calcFactorial(int num) throws Exception {
         if (num < 0) {
             throw new Exception("Non positive value");
         }
@@ -32,16 +33,16 @@ public class FactorialUtils {
      * @return возвращает значение, расчитанное по формуле
      * @throws Exception
      */
-    public static double calcF(int m, int r) throws Exception {
+    public static long calcF(int m, int r) throws Exception {
         int a = Integer.min(m - r, r);
         int b = Integer.max(m - r, r);
 
         //подсчет факториала для минимального значения из 3-х
-        int aFactorial = calcFactorial(a);
+        long aFactorial = calcFactorial(a);
         //подсчет факториала следующего значения, домножив первый найденный факториал на недостоющие значения
-        int bFactorial = IntStream.range(a + 1, b + 1).reduce(aFactorial, (res, value) -> res * value);
+        long bFactorial = LongStream.range(a + 1, b + 1).reduce(aFactorial, (res, value) -> res * value);
         //подсчет факториала следующего значения, домножив первый найденный факториал на недостоющие значения
-        int mFactorial = IntStream.range(b + 1, m + 1).reduce(bFactorial, (res, value) -> res * value);
+        long mFactorial = LongStream.range(b + 1, m + 1).reduce(bFactorial, (res, value) -> res * value);
 
         return mFactorial / (aFactorial * bFactorial);
     }
